+++ food-truck-api/src/main/java/food/truck/api/truck/Truck.java
@@ -5,11 +5,11 @@ public class Truck {
     private String name;
     private String description;
     private Float rating;
-    private Float price;
+    private Integer price;
     private String type;
     private String menu;
 
-    public Truck(String id, String name, String description, Float rating, String type, Float price) {
+    public Truck(String id, String name, String description, Float rating, String type, Integer price) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -18,7 +18,7 @@ public class Truck {
 		this.price = price;
         this.menu = "null";
     }
-    public Truck(String id, String name, String description, Float rating, String type, Float price, String menu) {
+    public Truck(String id, String name, String description, Float rating, String type, Integer price, String menu) {
         this.id = id;
         this.name = name;
         this.description = description;
